+++ CHANGELOG.md
@@ -51,6 +51,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Versions of Librsvg prior to 2.36.0 are no longer supported.
 - Versions of GDK prior to 2.0.0 are no longer supported.
 - Versions of Glib prior to 2.36.0 are no longer supported.
+- **Breaking**: The `Agnodeinfo_t.inleaf` field and its `ND_inleaf` accessor
+  have been removed.
+- **Breaking**: The `Agnodeinfo_t.outleaf` field and its `ND_outleaf` and
+  `GD_outleaf` accessors have been removed.
+- **Breaking**: The `Agraphinfo_t.has_sourcerank` field and its
+  `GD_has_sourcerank` accessor has been removed.
+- **Breaking**: The `Agraphinfo_t.has_sinkrank` field and its
+  `GD_has_sinkrank` accessor has been removed.
 
 ## [7.1.0] – 2023-01-21